Only exclude inactive validators in the next era.
While malicious validators should be banned until they get removed from
the validator set by the auction contract, inactive validators just need
to be excluded from proposing blocks in the next era. If they become
active again, it is fine to add them back to the leader sequence in the
era after that.

This change doesn't make a difference in mainnet anyway, because the
`auction_delay` is 1.