contract messages: remove proposal for using per transaction message log
The per transaction message log offered a good solution to allow
corelation of message ordering across multiple topics and contracts that
emit messages.
Unfortunately writing the log creates many entries in global state that
are hard to prune in the next block since the transaction hashes are
unique. The only solution to reliably prune old entries is to iterate
through all of the log keys which is a time consuming process. Because
the logs could have the potential to widen the tip we decided to abandon
this solution.
A different proposal will be added in a later commit.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>